Reduce install command load time
- Split lightweight formula checks and the base cask class from their
  larger implementation graphs.
- Load full formula and cask installers only when their paths are used.
- Keep deferred consumers explicit about their runtime dependencies.
- Cut warm no-op `command_load` from 89.05ms to 51.11ms on this base.
- Avoid loading 133 features and 79 cask paths during dispatch.
