+++ Library/Homebrew/cask/cask_base.rb
@@ -1,0 +1,8 @@
+# typed: strict
+# frozen_string_literal: true
+
+module Cask
+  class Cask
+    extend Forwardable
+  end
+end
