+++ Library/Homebrew/formula_installer.rb
@@ -13,6 +13,7 @@ require "install_renamed"
 require "sandbox"
 require "development_tools"
 require "cache_store"
+require "download_queue"
 require "linkage_checker"
 require "messages"
 require "cask/caskroom"
